# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/resource/{resourceId}/applications/review": {
      "put": {
        "operationId": "ResourceReviewController_reviewApplications",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewApplicationsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review completed"
          }
        },
        "summary": "Review resource applications",
        "tags": [
          "Resource Review"
        ]
      }
    },
    "/api/resource/{resourceId}/applications": {
      "get": {
        "operationId": "ResourceReviewController_getPendingApplications",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending applications retrieved successfully"
          }
        },
        "summary": "Get pending applications for a resource",
        "tags": [
          "Resource Review"
        ]
      }
    },
    "/api/resource/{condition_id}/edit": {
      "put": {
        "operationId": "ResourceConditionController_editCondition",
        "parameters": [
          {
            "name": "condition_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertResourceConditionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Condition updated successfully."
          },
          "404": {
            "description": "Condition not found."
          }
        },
        "summary": "Edit a resource condition by condition ID",
        "tags": [
          "Resource Condition"
        ]
      }
    },
    "/api/resource/{resource_id}/condition": {
      "post": {
        "operationId": "ResourceConditionController_createCondition",
        "parameters": [
          {
            "name": "resource_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertResourceConditionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Condition created successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "summary": "Create a new resource condition for a resource",
        "tags": [
          "Resource Condition"
        ]
      },
      "get": {
        "operationId": "ResourceConditionController_getConditions",
        "parameters": [
          {
            "name": "resource_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conditions retrieved successfully."
          }
        },
        "summary": "Get all conditions for a resource",
        "tags": [
          "Resource Condition"
        ]
      },
      "delete": {
        "operationId": "ResourceConditionController_deleteAllByResource",
        "parameters": [
          {
            "name": "resource_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All conditions deleted successfully."
          }
        },
        "summary": "Delete all conditions for a resource",
        "tags": [
          "Resource Condition"
        ]
      }
    },
    "/api/resource/condition/{condition_id}": {
      "delete": {
        "operationId": "ResourceConditionController_deleteCondition",
        "parameters": [
          {
            "name": "condition_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Condition deleted successfully."
          },
          "404": {
            "description": "Condition not found."
          }
        },
        "summary": "Delete a specific condition by condition ID",
        "tags": [
          "Resource Condition"
        ]
      }
    },
...
  },
  "info": {
    "title": "API docs",
    "description": "My NestJS API manual",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ReviewApplicationsDto": {
        "type": "object",
        "properties": {}
      },
      "UpsertResourceConditionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}